fix tests after update to main
