+++ app/assess/templates/assessor_tasklist.html
@@ -16,11 +16,10 @@
         </p>
         {{ logout_partial(sso_logout_url) }}
     </div>
-
-    {% if flag and g.user.highest_role in ("ASSESSOR", "LEAD_ASSESSOR") and flag.flag_type.name == "FLAGGED" %}
-        {{ assessment_flagged(flag, flag_user_info, application_id) }}
-    {% elif flag and g.user.highest_role in ("ASSESSOR", "LEAD_ASSESSOR") and flag.flag_type.name == "STOPPED" %}
-        {{ assessment_stopped(flag, flag_user_info, application_id) }}
+    {% if flag and current_user_role in ("ASSESSOR", "LEAD_ASSESSOR") and flag.flag_type.name == "FLAGGED" %}
+        {{ assessment_flagged(flag, flag_user_info, application_id, current_user_role) }}
+    {% elif flag and current_user_role in ("ASSESSOR", "LEAD_ASSESSOR") and flag.flag_type.name == "STOPPED" %}
+        {{ assessment_stopped(flag, flag_user_info, application_id, current_user_role) }}
     {% endif %}
 
     {{ banner_summary(
